fix: fixed package-lock and package.json conflict (temporary).
moved react-datepicker and react-day-picker to devDependencies.

moved type Product in types/product
