+++ mock_data/dummyData.ts
@@ -1,15 +1,3 @@
-export type Product = {
-  id: number;
-  name: string;
-  category: string;
-  description: string;
-  contactPerson: string;
-  address: string;
-  contactNumber: string;
-  productGallery: string[];
-  majorProductLines: string[];
-};
-
 export const sectors = [
   "All",
   "Bamboo",
