+++ package.json
@@ -9,7 +9,7 @@
     "db:migrate": "prisma migrate deploy",
     "db:push": "prisma db push",
     "db:studio": "prisma studio",
-    "dev": "next dev --turbo",
+    "dev": "next dev",
     "lint": "next lint",
     "start": "next start",
     "prepare": "husky"
@@ -54,11 +54,10 @@
     "geist": "^1.3.0",
     "input-otp": "^1.2.4",
     "lucide-react": "^0.453.0",
-    "next": "15.0.1-canary.1",
+    "next": "^14.2.4",
     "next-themes": "^0.3.0",
-    "react": "19.0.0-rc-69d4b800-20241021",
-    "react-day-picker": "^8.10.1",
-    "react-dom": "19.0.0-rc-69d4b800-20241021",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.0",
     "react-resizable-panels": "^2.1.4",
     "recharts": "^2.13.0",
@@ -72,26 +71,24 @@
   "devDependencies": {
     "@types/eslint": "^8.56.10",
     "@types/node": "^20.14.10",
-    "@types/react": "npm:types-react@19.0.0-rc.1",
-    "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@typescript-eslint/eslint-plugin": "^8.1.0",
     "@typescript-eslint/parser": "^8.1.0",
     "eslint": "^8.57.0",
-    "eslint-config-next": "15.0.1-canary.1",
+    "eslint-config-next": "^14.2.4",
     "husky": "^9.1.6",
     "lint-staged": "^15.2.10",
     "postcss": "^8.4.47",
     "prettier": "^3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
     "prisma": "^5.14.0",
+    "react-datepicker": "^7.5.0",
+    "react-day-picker": "^9.2.0",
     "tailwindcss": "^3.4.3",
     "typescript": "^5.5.3"
   },
   "ct3aMetadata": {
     "initVersion": "7.37.0"
-  },
-  "overrides": {
-    "@types/react": "npm:types-react@19.0.0-rc.1",
-    "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1"
   }
 }
