+++ src/types/products.ts
@@ -1,0 +1,11 @@
+export type Product = {
+  id: number;
+  name: string;
+  category: string;
+  description: string;
+  contactPerson: string;
+  address: string;
+  contactNumber: string;
+  productGallery: string[];
+  majorProductLines: string[];
+};
